When user registers will now convert the initial balance to desired currency

+++ src/main/java/dao/Account/AccountDao.java
@@ -4,5 +4,5 @@ import dao.DAO;
 import entity.Account;
 
 public interface AccountDao extends DAO<Account, Long> {
-    void updateBalance(int newBalance, long id);
+    void updateBalance(double newBalance, long id);
 }
